requirements.txt
pandas
numpy
requests
pyyaml
matplotlib
seaborn
scikit-learn
mlflow
dvc
reportlab
psycopg2